✨ refactor: Enhance syntax highlighting logic and cleanup
Update the `SyntaxHighlightAnnotation` to improve the handling of
text attributes for syntax highlighting. Introduce a null check for
`textAttributes` to enforce attributes correctly. Additionally, 
refactor the conditional logic for color handling, ensuring
proper usage of background color from `textAttributes`. 

Remove the unused `colorMap` to streamline the codebase and 
remove unnecessary clutter in the `Utils` file.
