+++ src/main/resources/messages/EditorGroupsBundle.properties
@@ -72,13 +72,13 @@ group.krasa.editorGroups.Actions.description=Editor groups menu
 group.krasa.editorGroups.Actions.text=Editor Groups
 hide.empty.groups=Hide Empty Groups
 hide.panel=Hide panel
-keyword.group.color=Tab background color
-keyword.group.disable=Disable indexing of this file
-keyword.group.fgcolor=Tab foreground color
-keyword.group.id=ID - optional for one group per file
-keyword.group.related=Related files
-keyword.group.root=Directory root, this by default
-keyword.group.title=Group title
+keyword.group.color=Group Tab background color
+keyword.group.disable=Disable Group
+keyword.group.fgcolor=Group Tab Foreground Color — Optional
+keyword.group.id=Group ID — Optional
+keyword.group.related=Group Related Files
+keyword.group.root=Group Root — Default to this
+keyword.group.title=Group Title
 krasa.editorGroups.EditorGroupsSettings=Editor Groups
 label.editorGroups.config.file=Editor groups custom group
 live.template.@.gc.description=Group color - EditorGroups plugin
